Stripe Payments Basics (#109)
* Cart updates

* Formatted cart checkout button correctly

* Further work on cart

* Added better pricing strings

* Updated to cart

* Enabled deleting from cart

* Started the basics of checkout with paypal

* Added better textfields for shop and fixed conflicting constraints

* Enabled add to cart during editing

* Integrated PayPal checkout with sandbox env

* Initial commit for second attempt at stripe

* More work with Stripe and node

* Created basic functionality to accept payments using Stripe in app

+++ gthockey-iOS/gthockey-iOS/JSONParser.swift
@@ -183,7 +183,7 @@ class JSONParser {
     private func makeApparelObject(value: JSON) -> Apparel {
         let apparel = Apparel(id: value["id"].int!,
                               name: value["name"].string!,
-                              price: value["price"].float!,
+                              price: value["price"].double!,
                               description: value["description"].string!,
                               imageURL: URL(string: value["image"].string ?? "https://test.gthockey.com/media/players/caleb.jpg")!)
         return apparel
@@ -209,7 +209,7 @@ class JSONParser {
                                                       displayName: item["display_name"].string!,
                                                       helpText: item["help_text"].string!,
                                                       isRequired: item["required"].bool!,
-                                                      extraCost: item["extra_cost"].int!,
+                                                      extraCost: item["extra_cost"].double!,
                                                       correspondingApparelID: item["shop_item"].int!)
             arr.append(apparelCustomItem)
         }
